+++ src/names.rs
@@ -1,23 +1,23 @@
 //! This module contains utilities to work with Tiled names.
 use bevy::{prelude::*, utils::HashSet};
 
-/// A struct to specify names when using [ObjectNameFilter]
+/// A struct to specify names when using [TiledNameFilter]
 ///
 /// Example:
 /// ```rust,no_run
 /// use bevy_ecs_tiled::prelude::*;
 ///
 /// let name_to_check = String::from("some name");
-/// let matching_allowed_names = ObjectNames::Names(vec!("some name".to_string()));
-/// let non_matching_allowed_names = ObjectNames::Names(vec!("some other name".to_string()));
+/// let matching_allowed_names = TiledName::Names(vec!("some name".to_string()));
+/// let non_matching_allowed_names = TiledName::Names(vec!("some other name".to_string()));
 ///
-/// assert_eq!(ObjectNameFilter::from(&matching_allowed_names).contains(&name_to_check), true);
-/// assert_eq!(ObjectNameFilter::from(&non_matching_allowed_names).contains(&name_to_check), false);
-/// assert_eq!(ObjectNameFilter::from(&ObjectNames::All).contains(&name_to_check), true);
-/// assert_eq!(ObjectNameFilter::from(&ObjectNames::None).contains(&name_to_check), false);
+/// assert_eq!(TiledNameFilter::from(&matching_allowed_names).contains(&name_to_check), true);
+/// assert_eq!(TiledNameFilter::from(&non_matching_allowed_names).contains(&name_to_check), false);
+/// assert_eq!(TiledNameFilter::from(&TiledName::All).contains(&name_to_check), true);
+/// assert_eq!(TiledNameFilter::from(&TiledName::None).contains(&name_to_check), false);
 /// ```
-#[derive(Default, Clone, PartialEq, Reflect)]
-pub enum ObjectNames {
+#[derive(Default, Clone, PartialEq, Reflect, Debug)]
+pub enum TiledName {
     /// Matches all names.
     #[default]
     All,
@@ -29,23 +29,23 @@ pub enum ObjectNames {
     None,
 }
 
-/// Allow to check if a provided [ObjectNames] matches a given name.
+/// Allow to check if a provided [TiledName] matches a given name.
 ///
 /// Example:
 /// ```rust,no_run
 /// use bevy_ecs_tiled::prelude::*;
 ///
 /// let name_to_check = String::from("some name");
-/// let matching_allowed_names = ObjectNames::Names(vec!("some name".to_string()));
-/// let non_matching_allowed_names = ObjectNames::Names(vec!("some other name".to_string()));
+/// let matching_allowed_names = TiledName::Names(vec!("some name".to_string()));
+/// let non_matching_allowed_names = TiledName::Names(vec!("some other name".to_string()));
 ///
-/// assert_eq!(ObjectNameFilter::from(&matching_allowed_names).contains(&name_to_check), true);
-/// assert_eq!(ObjectNameFilter::from(&non_matching_allowed_names).contains(&name_to_check), false);
-/// assert_eq!(ObjectNameFilter::from(&ObjectNames::All).contains(&name_to_check), true);
-/// assert_eq!(ObjectNameFilter::from(&ObjectNames::None).contains(&name_to_check), false);
+/// assert_eq!(TiledNameFilter::from(&matching_allowed_names).contains(&name_to_check), true);
+/// assert_eq!(TiledNameFilter::from(&non_matching_allowed_names).contains(&name_to_check), false);
+/// assert_eq!(TiledNameFilter::from(&TiledName::All).contains(&name_to_check), true);
+/// assert_eq!(TiledNameFilter::from(&TiledName::None).contains(&name_to_check), false);
 /// ```
-#[derive(Clone)]
-pub enum ObjectNameFilter {
+#[derive(Clone, Debug)]
+pub enum TiledNameFilter {
     /// Matches all names.
     All,
     /// Matches only provided names.
@@ -54,31 +54,31 @@ pub enum ObjectNameFilter {
     None,
 }
 
-impl From<&ObjectNames> for ObjectNameFilter {
-    /// Initialize a [ObjectNameFilter] from an [ObjectNames].
-    fn from(value: &ObjectNames) -> Self {
+impl From<&TiledName> for TiledNameFilter {
+    /// Initialize a [TiledNameFilter] from an [TiledName].
+    fn from(value: &TiledName) -> Self {
         match value {
-            ObjectNames::All => ObjectNameFilter::All,
-            ObjectNames::Names(names) => {
+            TiledName::All => TiledNameFilter::All,
+            TiledName::Names(names) => {
                 let names = names
                     .iter()
                     .map(|x| x.trim().to_lowercase())
                     .filter(|x| !x.is_empty())
                     .collect();
-                ObjectNameFilter::Names(names)
+                TiledNameFilter::Names(names)
             }
-            ObjectNames::None => ObjectNameFilter::None,
+            TiledName::None => TiledNameFilter::None,
         }
     }
 }
 
-impl ObjectNameFilter {
+impl TiledNameFilter {
     /// Determine if provided [str] matches the filter.
     pub fn contains(&self, name: &str) -> bool {
         match self {
-            ObjectNameFilter::All => true,
-            ObjectNameFilter::Names(names) => names.contains(name),
-            ObjectNameFilter::None => false,
+            TiledNameFilter::All => true,
+            TiledNameFilter::Names(names) => names.contains(name),
+            TiledNameFilter::None => false,
         }
     }
 }